ch4/rma: fix accumulate compute with packed pairtype
When source buffer is not predefined, SHM RMA internally packs the data
before performing accumulate computation. If the basic datatype is a
pairtype, then the packed elements are strided by size but not extent.
Thus, we should use size as the source stride during computation. If the
source buffer is not packed (e.g., in active message), we still use extent
as stride.