ch4/ipc: rename param of MPIDI_GPU_ipc_handle_map
MPIDI_GPU_ipc_handle_map and its subroutine get_map_device involve
several parameters for GPU devices including the input `dev_handle`
as recv device in pt2pt case, `dev_id` as the device id which will
be used for mapping (note that `dev_id` is decided by get_map_device
but it might not be the id of input `dev_handle`), reuse of `dev_handle`
as the corresponding handle of `dev_id`. The naming of these parameters
does not match the usage. This patch renames these parameters to make
the purpose of each parameter clear.

As the set of functions will be used for both pt2pt and rma, the new
names follow local/remote* as to the process calling IPC mapping of
remote ptr. We rename recv_type->local_type for consistency.